you can pass smart pointers straight into API functions without using .get()
git-svn-id: http://audiere.svn.sourceforge.net/svnroot/audiere/trunk@477 cb5fa4c3-b70d-0410-b4f6-90f5898e5689

+++ src/device_ds_buffer.cpp
@@ -100,7 +100,7 @@ namespace audiere {
 
   void
   DSOutputBuffer::setPitchShift(float shift) {
-    m_buffer->SetFrequency(m_base_frequency * shift);
+    m_buffer->SetFrequency(DWORD(m_base_frequency * shift));
   }
